+++ src/components/banner/banner-image-container/BannerImageContainer.tsx
