+++ src/stories/components/banner/BannerGradient.stories.tsx
@@ -1,10 +1,10 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { BannerGradient } from '@/components/banner-section/banner-gradient/BannerGradient';
+import { BannerGradient } from '@/components/banner/banner-gradient/BannerGradient';
 
 const meta: Meta<typeof BannerGradient> = {
   component: BannerGradient,
-  title: 'Components/BannerSection/BannerGradient',
+  title: 'Components/Banner/BannerGradient',
   argTypes: {},
 };
